panel and dashboard migration combined
