+++ web/src/utils/dashboard/convertDashboardSchemaVersion.ts
@@ -12,7 +12,51 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-import { convertPanelSchemaVersion } from "./convertPanelSchemaVersion";
+const convertPanelSchemaVersion = (data: any) => {
+  if (!data || (typeof data === "object" && Object.keys(data).length === 0)) {
+    return;
+  }
+  if (!data.version) data = { ...data, version: 1 };
+  switch (data.version) {
+    case 1: {
+      // converting this to new array as z axis is added in the heatmap
+      const queryFields = {
+        stream_type: data.fields?.stream_type || "logs",
+        stream: data.fields.stream || "",
+        x: data.fields?.x || [],
+        y: data.fields?.y || [],
+        z: [], // this is a new field
+        filter: data.fields?.filter || [],
+      };
+
+      data = {
+        version: 2,
+        id: data.id,
+        type: data.type,
+        config: {
+          title: data.config.title,
+          description: data.config.description,
+          show_legends: data.config.show_legends,
+          legends_position: data.config.legends_position,
+          unit: data.config.unit,
+          unit_custom: data.config.unit_custom,
+        },
+        queryType: data.queryType,
+        queries: [
+          {
+            query: data.query,
+            customQuery: data.customQuery,
+            fields: queryFields,
+            config: {
+              promql_legend: data.config.promql_legend,
+            },
+          },
+        ],
+      };
+    }
+  }
+  return data;
+}
 
 export function convertDashboardSchemaVersion(data: any) {
   if (!data) {
@@ -36,6 +80,7 @@ export function convertDashboardSchemaVersion(data: any) {
 
       // add layout object in panels array and also migrate panels schema using panelschemaversionconverted function
       data.panels = data?.panels?.map((panelItem: any) => ({
+        //
         ...convertPanelSchemaVersion(panelItem),
         layout: layoutsObjBasedOnPanelId[panelItem.id], // Use the layout item from the mapping
       }));
@@ -52,3 +97,61 @@ export function convertDashboardSchemaVersion(data: any) {
   // return converted data
   return data;
 }
+
+
+
+// const dataV1 = {
+//   version: 1,
+//   id: "123",
+//   type: "bar",
+//   fields: {
+//     stream: "",
+//     stream_type: "logs",
+//     x: [],
+//     y: [],
+//     filter: [],
+//   },
+//   config: {
+//     title: "",
+//     description: "",
+//     show_legends: true,
+//     legends_position: null,
+//     promql_legend: "",
+//     unit: null,
+//     unit_custom: null,
+//   },
+//   queryType: "sql",
+//   query: "",
+//   customQuery: false,
+// };
+
+// const dataV2 = {
+//   version: 2,
+//   id: "456",
+//   type: "bar",
+//   config: {
+//     title: "",
+//     description: "",
+//     show_legends: true,
+//     legends_position: null,
+//     unit: null,
+//     unit_custom: null,
+//   },
+//   queryType: "sql",
+//   queries: [
+//     {
+//       query: "",
+//       customQuery: false,
+//       fields: {
+//         stream: "",
+//         stream_type: "logs",
+//         x: [],
+//         y: [],
+//         filter: [],
+//       },
+//       config: {
+//         promql_legend: "",
+//       },
+//     },
+//   ],
+// };
